Item Events
Added in some events to add and remove items from the inventory (at
chest and well). Need to fix the well and sign events as they don't
currently trigger - go back to the old method of having an 'invisible'
event on the tile before. Kind of wasteful in terms of memory, though.