Event Fixed and Readme Update
Corrected the events at the well, sign and mermaid statue. Now located
on the tile segment below the object itself, and activates on the
condition that the player is facing 'up' (towards the object). Does mean
that events don't activate from other angles, but do at least activate
in the first place.
Also updated the readme with better phrasing and correct map names.