Map Changes
Extended the EXT:IslandMap to be larger so that I could add more
buildings to the town and give it a less isolated feel. Starting to add
details for future events (graveyard, vines). Also added a new map,
INT:Manor, for the largest town building with placeholder tiles. Added
and fixed events for moving between maps, as well as amended Erethin's
dialogue (removed Vitali's response, felt it was too clunky).