Double Conditional Branch
Added an event to INT:Manor where different speech is shown depending on
whether the player has previously entered the house or not. As the event
is triggered when the player touches the event, a second branch was
needed to determine whether or not the event had been activated yet (the
speech would not be displayed if the event had already been triggered).
However, the event needed to play every time the house was exited and
re-entered. This commit contains the successful implementation of the
dual conditions.