Cosmetic Events
Added several events to the Manor maps, as well as one at a gravestone
in EXT:IslandMap. Most are for cosmetic purposes (descriptive text,
etc), though one event at the child's bed gives 1x Runed Robe.
Idea: The child who lives in the house has disappeared (following
Erethin's concerns) and the player must find her. The robe belongs to
her and offers great protective power. This will be the only time the
Runed Robe can be found or used in the game.