Dungeon Changes & Ella
Edited the DUN:LostTemple map to a more complete state - could possibly
be used as the 'end' room after the boss, housing Ella. Following that
idea, an 'NPC' Ella exists in the room. Need to add floors to the
dungeon and a way out (exit near the town?). Also need to generate the
event(s) to add Ella to the party.