Dungeon Expansion
Added two new levels (maps) to the dungeon, DUN:1F and DUN:B1F,
including an almost-identical copy with minor changes (DUN:1FP2). Also
added events to move between maps depending on whether or not the
'puzzle' had been solved. The puzzle involves pushing a boulder onto the
correct floor tile.

Need to learn how to script so that the push event can be made - for
now, the event is triggered by the action button on the finish tile.
Added a few dialogue events and edited Ella's event so that she is now
added to the party after speaking to her.