Enemy Encounters
Returned to this project after a brief hiatus to add in enemy encounters
to the dungeon maps. Added two encounters - bat x2 and spider x3 - to
the first dungeon room (DUN:1F and DUN:1FP2). Experimented with the
rarity of monsters to find a balance where bats are mostly encountered
with some spiders here and there. Also changed some of Vitali's equipped
items in order to balance the game a little better.