DUN:B1F Expansion and Event Errors
Expanded the DUN:B1F map to include a "jumping puzzle" requiring players
to jump over gaps in the floor. Need to actually add the tags for "void"
tiles, although some "action button" events have stopped working,
notably those in Ella's Bedroom which give access to the dungeon. It
could be a bug, though it is more likely that something was set up
incorrectly.

TODO:
- Complete implementation of jumping puzzle in dungeon.
- Add battles to DUN:B1F
- Fix action button events (possibly remake Ella's Room map?)