Script Usage Begun
I began testing and using the Victor Engine scripts more thoroughly,
discovering that the 'bug' that was causing problems with action button
events was actually just the Pixel Movement script changing where events
are detected (i.e. it's now at the bottom of an event). I will adjust
this to be more centred if I can.

Also added the map notes and comment calls to use the Void Terrain
element of the Moving Platform script. This is not working as yet, but
I'm not sure what I've missed.